Split message module, reorganise some modules
- Separate message module into type module, background-messaging, and
  tab-messaging
- Rename "highlight messages" to "tab messages"
- Move modules requiring background privilege into modules/privilege/
