+++ src/modules/interface/terms-setter-legacy.mts
@@ -1,13 +1,11 @@
-import {
-	type HighlightMessage, type HighlightMessageResponse,
-	messageSendBackground,
-} from "/dist/modules/message.mjs";
+import type * as Message from "/dist/modules/messaging.mjs";
+import { sendBackgroundMessage } from "/dist/modules/messaging/background.mjs";
 import { MatchTerm } from "/dist/modules/match-term.mjs";
 
 type MessageHandler = (
-	message: HighlightMessage,
+	message: Message.Tab,
 	sender: chrome.runtime.MessageSender | null,
-	sendResponse: (response: HighlightMessageResponse) => void,
+	sendResponse: (response: Message.TabResponse) => void,
 ) => void
 
 let messageHandleHighlightGlobal: MessageHandler = () => undefined;
@@ -18,7 +16,7 @@ const registerMessageHandler = (handler: MessageHandler) => {
 
 const termsSet = async (terms: Array<MatchTerm>) => {
 	messageHandleHighlightGlobal({ terms: terms.slice() }, null, () => undefined);
-	await messageSendBackground({ terms });
+	await sendBackgroundMessage({ terms });
 };
 
 export { registerMessageHandler, termsSet };
