+++ src/modules/messaging/background.mts
@@ -1,0 +1,8 @@
+import type * as Message from "/dist/modules/messaging.mjs";
+
+// TODO document
+const sendBackgroundMessage = (message: Message.Background): Promise<Message.BackgroundResponse> =>
+	chrome.runtime.sendMessage(message)
+;
+
+export { sendBackgroundMessage };
